Refactored AboutWindow
